requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.6
django-decouple==2.1
django-heroku==0.3.1
djangorestframework==3.12.2
gunicorn==20.0.4
importlib-metadata==3.4.0
Markdown==3.3.3
psycopg2==2.8.6
Pygments==2.7.4
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
whitenoise==5.2.0
zipp==3.4.0
